fix warning code, update coin question status when the status question is blank state

+++ MarioGame/Animations.cpp
@@ -20,7 +20,7 @@ void CAnimation::Add(int spriteId, DWORD time)
 }
 
 // NOTE: sometimes Animation object is NULL ??? HOW ??? 
-void CAnimation::Render(float x, float y, int alpha)
+void CAnimation::Render(double x, double y, int alpha)
 {
 	DWORD now = GetTickCount();
 	if (currentFrame == -1)
