+++ MarioGame/BackgroundObject.h
@@ -6,5 +6,5 @@ class CBackgroundObject : public CGameObject
 {
 public:
 	virtual void Render();
-	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
+	virtual void GetBoundingBox(double& l, double& t, double& r, double& b);
 };
