+++ MarioGame/FireFlower.h
@@ -15,7 +15,7 @@ public:
 	virtual void Render();
 	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
 	virtual void SetState(int state);
-	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
+	virtual void GetBoundingBox(double& l, double& t, double& r, double& b);
 	void SetIsFiring(bool value) { isFiring = value; };
 
 	bool GetIsFiring() { return isFiring; }
