+++ MarioGame/Flower.h
@@ -17,7 +17,7 @@ private:
 public:
 	CFlower();
 	virtual void Render();
-	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
+	virtual void GetBoundingBox(double& l, double& t, double& r, double& b);
 	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
 	virtual void SetState(int state);
 	void SetHasFired(bool value) { hasFired = value; };
