+++ MarioGame/Koopas.cpp
@@ -13,7 +13,7 @@ CKoopas::CKoopas()
 
 }
 
-void CKoopas::GetBoundingBox(float& left, float& top, float& right, float& bottom)
+void CKoopas::GetBoundingBox(double& left, double& top, double& right, double& bottom)
 {
 	left = x;
 	top = y;
@@ -87,9 +87,9 @@ void CKoopas::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
 	}
 	else
 	{
-		float min_tx, min_ty, nx = 0, ny;
-		float rdx = 0;
-		float rdy = 0;
+		double min_tx, min_ty, nx = 0, ny;
+		double rdx = 0;
+		double rdy = 0;
 
 		collisionHandler->FilterCollision(coEvents, coEventsResult, min_tx, min_ty, nx, ny, rdx, rdy);
